move utilities to utils/sentry_utils.py
